pyproject.toml
[project]
name = "limitless-lifelog"
version = "0.1.0"
description = "CLI agent for processing Limitless Voice API life logs and integrating with Notion"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.49.0",
    "openai>=1.68.0",
    "python-dotenv>=1.0.1",
    "pydantic>=2.0.0",
    "mcp>=0.1.5",
    "requests>=2.31.0",
    "notion-client>=2.0.0",
    "argparse>=1.4.0",
    "loguru>=0.7.0",
]

[project.scripts]
lifelog = "limitless_lifelog.__main__:main"

[project.optional-dependencies]
test = [
    "pytest>=7.3.1",
    "pytest-asyncio>=0.20.3",
    "pytest-mock>=3.10.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"